Spline Based Room Generator
Found an awesome function that handles mesh triangulation from vertices called PolygonX.
Used the spline points as vertices for a procedural mesh and then triangulated those with PolygonX, to create the floor and ceiling meshes. 
Walls are currently cubes (Will be modeled+textured by end of sprint) that are placed as spline mesh components along spline.